requirements.txt
bcrypt==3.2.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
docutils==0.17.1
email-validator==1.1.2
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
google-api-core==1.28.0
google-auth==1.30.1
google-cloud-language==2.0.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.38.0
idna==2.10
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
libcst==0.3.19
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.15
statistics==1.0.3.5
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.4.1
